+++ catalog/app/containers/Admin/Settings/TabulatorSettings.tsx
@@ -9,8 +9,8 @@ import * as Notifications from 'containers/Notifications'
 import * as GQL from 'utils/GraphQL'
 import StyledLink from 'utils/StyledLink'
 
-import UNRESTRICTED_QUERY from './gql/TabulatorUnrestricted.generated'
-import SET_UNRESTRICTED_MUTATION from './gql/SetTabulatorUnrestricted.generated'
+import OPEN_QUERY_QUERY from './gql/TabulatorOpenQuery.generated'
+import SET_OPEN_QUERY_MUTATION from './gql/SetTabulatorOpenQuery.generated'
 
 interface ToggleProps {
   checked: boolean
@@ -18,15 +18,15 @@ interface ToggleProps {
 
 function Toggle({ checked }: ToggleProps) {
   const { push: notify } = Notifications.use()
-  const mutate = GQL.useMutation(SET_UNRESTRICTED_MUTATION)
-  const [mutation, setMutation] = React.useState<{ value: boolean } | null>(null)
+  const mutate = GQL.useMutation(SET_OPEN_QUERY_MUTATION)
+  const [mutation, setMutation] = React.useState<{ enabled: boolean } | null>(null)
 
   const handleChange = React.useCallback(
-    async (_event, value: boolean) => {
+    async (_event, enabled: boolean) => {
       if (mutation) return
-      setMutation({ value })
+      setMutation({ enabled })
       try {
-        await mutate({ value })
+        await mutate({ enabled })
       } catch (e) {
         Sentry.captureException(e)
         notify(`Failed to update tabulator settings: ${e}`)
@@ -42,18 +42,18 @@ function Toggle({ checked }: ToggleProps) {
       <M.FormControlLabel
         control={
           <M.Switch
-            checked={mutation?.value ?? checked}
+            checked={mutation?.enabled ?? checked}
             onChange={handleChange}
             disabled={!!mutation}
           />
         }
-        label="Enable unrestricted access"
+        label="Enable open querying of Tabulator tables"
       />
       <M.FormHelperText>
         <b>CAUTION:</b> When enabled, Tabulator defers all access control to AWS and does
         not enforce any extra restrictions.{' '}
         <StyledLink
-          href={`${docs}/advanced-features/tabulator#unrestricted-access`}
+          href={`${docs}/advanced-features/tabulator#open-query`}
           target="_blank"
         >
           Learn more
@@ -65,12 +65,12 @@ function Toggle({ checked }: ToggleProps) {
 }
 
 export default function TabulatorSettings() {
-  const query = GQL.useQuery(UNRESTRICTED_QUERY)
+  const query = GQL.useQuery(OPEN_QUERY_QUERY)
 
   return (
     <M.FormGroup>
       {GQL.fold(query, {
-        data: ({ admin }) => <Toggle checked={admin.tabulatorUnrestricted} />,
+        data: ({ admin }) => <Toggle checked={admin.tabulatorOpenQuery} />,
         fetching: () => (
           <>
             <Skeleton width="40%" height={38} />